Use simple lambda function to show pass through
According to [this function](https://github.com/mlrun/mlrun/blob/1f15e9b37801f2902ffcd14a5b096ef6a8006d60/mlrun/utils/helpers.py#L833), it is perfectly valid to pass in a callable directly. I believe this is more clear and explicit than specifying a string that is mapped into a lambda. It should probably also be in the documentation what the _fn is with respect to storey UnaryFunctionFlow and how that is used for the different transformation flows.