[Artifacts] Fix Artifact Deletion by UID [1.9.x] (#8193)
When deleting an artifact, we expect it to be deleted by its UID.
However, `manager.delete_artifact` was not passing `item.uid` to
`httpdb.del_artifact`. As a result, when attempting to delete an
artifact with the same key and project but with a different UID, it
would delete all artifacts matching the key and project, ignoring the
UID.

To fix this, I passed the UID explicitly. I also removed the tag
parameter, since the db layer does not support receiving both tag and
uid simultaneously. Besides, we're deleting a specific artifact, not an
artifact tag.
https://iguazio.atlassian.net/browse/ML-10296

(cherry picked from commit b1ac3bda3975432c0ad5e4addeeaaf051d17b2a2)